# backend-api.openapi.yaml
openapi: 3.0.3
info:
  title: Car Service Admin API
  version: 2.0.0
  description: |
    Backend contract for the Car Service Admin dashboard.

    The frontend currently covers authentication, onboarding, dashboard widgets,
    bookings, payments, payment exports, reports, storage inventory, business
    settings, billing account verification, notification preferences, inbox,
    feedback, help center content, and location search.

    API JSON uses camelCase. The recommended database schema uses snake_case.
    Dates should be stored and exchanged as ISO-8601 strings. The frontend can
    format those dates for display.
  contact:
    name: Car Service Admin
servers:
  - url: http://localhost:3333
    description: Local backend development
  - url: https://api.car-service-admin.example.com
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /business/profile:
    get:
      tags: [Business]
      operationId: getBusinessProfile
      summary: Get current business details and shop location
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    patch:
      tags: [Business]
      operationId: updateBusinessProfile
      summary: Update business details or shop location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUpdateDto"
      responses:
        "200":
          description: Updated business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /dashboard/bookings/month:
    get:
      tags: [Dashboard]
      operationId: getDashboardMonthBookings
      summary: Get bookings that overlap a calendar month
      parameters:
        - in: query
          name: year
          required: true
          schema:
            type: integer
            minimum: 2000
            example: 2026
        - in: query
          name: month
          required: true
          description: Zero-based month number. 0 is January.
          schema:
            type: integer
            minimum: 0
            maximum: 11
            example: 3
      responses:
        "200":
          description: Calendar month bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarBookings"
  /dashboard/widgets:
    get:
      tags: [Dashboard]
      operationId: getDashboardWidgets
      summary: Get active and available dashboard widgets
      responses:
        "200":
          description: Dashboard widget preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardWidgetPreferences"
    put:
      tags: [Dashboard]
      operationId: updateDashboardWidgets
      summary: Replace active dashboard widget order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardWidgetPreferencesUpdateDto"
      responses:
        "200":
          description: Updated dashboard widget preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardWidgetPreferences"
...
  /customers:
    get:
      tags: [Customers]
      operationId: listCustomers
      summary: List customers for the current business
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Paginated customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCustomers"
    post:
      tags: [Customers]
      operationId: createCustomer
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerCreateDto"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/ValidationError"
  /customers/{customerId}:
    get:
      tags: [Customers]
      operationId: getCustomer
      summary: Get customer details
      parameters:
        - $ref: "#/components/parameters/CustomerIdPath"
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Customers]
      operationId: updateCustomer
      summary: Update customer details
      parameters:
        - $ref: "#/components/parameters/CustomerIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdateDto"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /business/services:
    get:
      tags: [Services]
      operationId: listBusinessServices
      summary: List services offered by the current business
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/ServiceStatus"
      responses:
        "200":
          description: Business services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessService"
    post:
      tags: [Services]
      operationId: createBusinessService
      summary: Add a service to the current business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessServiceCreateDto"
      responses:
        "201":
          description: Created business service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessService"
        "400":
          $ref: "#/components/responses/ValidationError"
  /business/services/{businessServiceId}:
    patch:
      tags: [Services]
      operationId: updateBusinessService
      summary: Update a business service price, category, name, or status
      parameters:
        - $ref: "#/components/parameters/BusinessServiceIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessServiceUpdateDto"
      responses:
        "200":
          description: Updated business service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessService"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Services]
      operationId: deleteBusinessService
      summary: Remove a business service
      parameters:
        - $ref: "#/components/parameters/BusinessServiceIdPath"
      responses:
        "204":
          description: Removed
        "404":
          $ref: "#/components/responses/NotFound"
...
  /inventory/items:
    get:
      tags: [Inventory]
      operationId: listInventoryItems
      summary: List inventory items/products with filters and pagination
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: brand
          description: Alias for company.
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: viscosity
          schema:
            type: string
        - in: query
          name: volume
          schema:
            type: string
        - in: query
          name: application
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Paginated inventory items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedInventoryItems"
    post:
      tags: [Inventory]
      operationId: createInventoryItem
      summary: Create an inventory item/product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemCreateDto"
      responses:
        "201":
          description: Created inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "400":
          $ref: "#/components/responses/ValidationError"
  /inventory/items/{inventoryItemId}:
    get:
      tags: [Inventory]
      operationId: getInventoryItem
      summary: Get inventory item details
      parameters:
        - $ref: "#/components/parameters/InventoryItemIdPath"
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Inventory]
      operationId: updateInventoryItem
      summary: Update an inventory item/product
      parameters:
        - $ref: "#/components/parameters/InventoryItemIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdateDto"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Inventory]
      operationId: deleteInventoryItem
      summary: Delete or archive an inventory item
      parameters:
        - $ref: "#/components/parameters/InventoryItemIdPath"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /inventory/items/{inventoryItemId}/sell:
    post:
      tags: [Inventory]
      operationId: sellInventoryItem
      summary: Sell an inventory item and create a payment/stock transaction
      parameters:
        - $ref: "#/components/parameters/InventoryItemIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellInventoryItemDto"
      responses:
        "201":
          description: Sale recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventorySaleResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /inventory/transactions:
    get:
      tags: [Inventory]
      operationId: listInventoryTransactions
      summary: List stock movements and item sale transactions
      parameters:
        - in: query
          name: inventoryItemId
          schema:
            type: string
            format: uuid
        - in: query
          name: type
          schema:
            $ref: "#/components/schemas/InventoryTransactionType"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Paginated stock transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedInventoryTransactions"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageQuery:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitQuery:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    CustomerIdPath:
      in: path
      name: customerId
      required: true
      schema:
        type: string
        format: uuid
...
    BusinessServiceIdPath:
      in: path
      name: businessServiceId
      required: true
      schema:
        type: string
        format: uuid
    InventoryItemIdPath:
      in: path
      name: inventoryItemId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      required: [statusCode, code, message]
      properties:
        statusCode:
          type: integer
          example: 400
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Please enter a valid email address.
        details:
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorDetail"
        requestId:
          type: string
          nullable: true
    ApiErrorDetail:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: Please enter a valid email address.
...
    Pagination:
      type: object
      required: [total, page, limit, pages]
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
...
    Address:
      type: object
      required: [country, streetAddress, city, state, zipCode]
      properties:
        country:
          type: string
          example: Saudi Arabia
        streetAddress:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        lat:
          type: number
          format: double
          nullable: true
        lng:
          type: number
          format: double
          nullable: true
...
    BusinessType:
      type: string
      enum: [general-repair, specialized, body-paint, quick-service]
    EmployeeCount:
      type: string
      enum: [small, medium, large]
...
    WorkingHours:
      type: object
      properties:
        open:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "09:00"
        close:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "18:00"
        timezone:
          type: string
          example: Africa/Cairo
    Business:
      type: object
      required: [id, name, email, phoneNumber, address, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        commercialRegistrationNumber:
          type: string
          nullable: true
        taxIdentificationNumber:
          type: string
          nullable: true
        businessType:
          $ref: "#/components/schemas/BusinessType"
        employeeCount:
          $ref: "#/components/schemas/EmployeeCount"
        address:
          $ref: "#/components/schemas/Address"
        workingHours:
          $ref: "#/components/schemas/WorkingHours"
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BusinessUpdateDto:
      type: object
      properties:
        name:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
          format: email
        commercialRegistrationNumber:
          type: string
        taxIdentificationNumber:
          type: string
        businessType:
          $ref: "#/components/schemas/BusinessType"
        employeeCount:
          $ref: "#/components/schemas/EmployeeCount"
        address:
          $ref: "#/components/schemas/Address"
        workingHours:
          $ref: "#/components/schemas/WorkingHours"
...
    CalendarBookings:
      type: object
      required: [year, month, bookings]
      properties:
        year:
          type: integer
        month:
          type: integer
          minimum: 0
          maximum: 11
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
    WidgetType:
      type: string
      enum: [profile, calendar, status, setupWarnings]
    DashboardWidget:
      type: object
      required: [id, title, type, span]
      properties:
        id:
          type: string
          example: setup-warnings
        title:
          type: string
        type:
          $ref: "#/components/schemas/WidgetType"
        statusKey:
          type: string
          enum: [inProgress, completed, pending, cancelled]
          nullable: true
        color:
          type: string
          nullable: true
          example: "#10b981"
        span:
          type: integer
          minimum: 1
          maximum: 4
    DashboardWidgetPreferences:
      type: object
      required: [activeWidgets, inactiveWidgets]
      properties:
        activeWidgets:
          type: array
          items:
            $ref: "#/components/schemas/DashboardWidget"
        inactiveWidgets:
          type: array
          items:
            $ref: "#/components/schemas/DashboardWidget"
    DashboardWidgetPreferencesUpdateDto:
      type: object
      required: [activeWidgetIds]
      properties:
        activeWidgetIds:
          type: array
          items:
            type: string
          example: [profile, calendar, setup-warnings, in-progress, completed, pending, cancelled]
...
    Customer:
      type: object
      required: [id, businessId, firstName, lastName, createdAt]
      properties:
        id:
          type: string
          format: uuid
        businessId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CustomerCreateDto:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        avatarUrl:
          type: string
    CustomerUpdateDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        avatarUrl:
          type: string
    PaginatedCustomers:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    ServiceStatus:
      type: string
      enum: [active, inactive]
    BusinessService:
      type: object
      required: [id, businessId, category, name, price, status]
      properties:
        id:
          type: string
          format: uuid
        businessId:
          type: string
          format: uuid
        serviceId:
          type: string
          nullable: true
          description: Seeded catalog service id when this row maps to catalog.
        category:
          type: string
          example: Engine Services
        name:
          type: string
          example: Oil Change
        price:
          type: number
          format: double
        currency:
          type: string
          default: USD
        status:
          $ref: "#/components/schemas/ServiceStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BusinessServiceCreateDto:
      type: object
      required: [category, name, price]
      properties:
        serviceId:
          type: string
          nullable: true
        category:
          type: string
        name:
          type: string
        price:
          type: number
          format: double
        currency:
          type: string
          default: USD
        status:
          $ref: "#/components/schemas/ServiceStatus"
    BusinessServiceUpdateDto:
      type: object
      properties:
        serviceId:
          type: string
          nullable: true
        category:
          type: string
        name:
          type: string
        price:
          type: number
          format: double
        currency:
          type: string
        status:
          $ref: "#/components/schemas/ServiceStatus"
...
    InventoryItemStatus:
      type: string
      enum: [active, inactive, archived]
    InventoryItem:
      type: object
      required: [id, businessId, category, name, price, quantity, status]
      properties:
        id:
          type: string
          format: uuid
        businessId:
          type: string
          format: uuid
        category:
          type: string
          example: Oil
        name:
          type: string
          example: Synthetic Engine Oil
        company:
          type: string
          nullable: true
          example: Shell
        brand:
          type: string
          nullable: true
          description: Display alias for company used by the storage inventory filters.
        type:
          type: string
          nullable: true
          example: Fully Synthetic
        specification:
          type: string
          nullable: true
          example: 5W-30
        viscosity:
          type: string
          nullable: true
          example: 5W-30
        capacity:
          type: string
          nullable: true
          example: 4L
        volume:
          type: string
          nullable: true
          example: 4L
        apiSpec:
          type: string
          nullable: true
          example: SN
        application:
          type: string
          nullable: true
          example: Gasoline
        details:
          type: string
          nullable: true
        usedInServices:
          type: array
          items:
            type: string
        imageUrl:
          type: string
          nullable: true
        price:
          type: number
          format: double
        purchasePrice:
          type: number
          format: double
          nullable: true
        currency:
          type: string
          default: USD
        deliveryTime:
          type: string
          nullable: true
          example: 2 Days
        quantity:
          type: integer
          minimum: 0
        lowStockThreshold:
          type: integer
          minimum: 0
          default: 0
        status:
          $ref: "#/components/schemas/InventoryItemStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InventoryItemCreateDto:
      type: object
      required: [category, name, price]
      properties:
        category:
          type: string
        name:
          type: string
        company:
          type: string
        brand:
          type: string
        type:
          type: string
        specification:
          type: string
        viscosity:
          type: string
        capacity:
          type: string
        volume:
          type: string
        apiSpec:
          type: string
        application:
          type: string
        details:
          type: string
        usedInServices:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        price:
          type: number
          format: double
        purchasePrice:
          type: number
          format: double
        currency:
          type: string
          default: USD
        deliveryTime:
          type: string
        quantity:
          type: integer
          minimum: 0
          default: 0
        lowStockThreshold:
          type: integer
          minimum: 0
          default: 0
        status:
          $ref: "#/components/schemas/InventoryItemStatus"
    InventoryItemUpdateDto:
      type: object
      properties:
        category:
          type: string
        name:
          type: string
        company:
          type: string
        brand:
          type: string
        type:
          type: string
        specification:
          type: string
        viscosity:
          type: string
        capacity:
          type: string
        volume:
          type: string
        apiSpec:
          type: string
        application:
          type: string
        details:
          type: string
        usedInServices:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        price:
          type: number
          format: double
        purchasePrice:
          type: number
          format: double
        currency:
          type: string
        deliveryTime:
          type: string
        quantity:
          type: integer
          minimum: 0
        lowStockThreshold:
          type: integer
          minimum: 0
        status:
          $ref: "#/components/schemas/InventoryItemStatus"
    PaginatedInventoryItems:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
    SellInventoryItemDto:
      type: object
      required: [quantity, paymentMethod]
      properties:
        quantity:
          type: integer
          minimum: 1
          default: 1
        customerId:
          type: string
          format: uuid
          nullable: true
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        unitPrice:
          type: number
          format: double
          nullable: true
    InventorySaleResult:
      type: object
      required: [item, transaction, payment]
      properties:
        item:
          $ref: "#/components/schemas/InventoryItem"
        transaction:
          $ref: "#/components/schemas/InventoryTransaction"
        payment:
          $ref: "#/components/schemas/Payment"
    InventoryTransactionType:
      type: string
      enum: [purchase, sale, adjustment, return]
...
    PaginatedInventoryTransactions:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InventoryTransaction"
        pagination:
          $ref: "#/components/schemas/Pagination"
    PaymentMethod:
      type: string
      enum: [VISA, Fawry, PayPal, Cash]
...